+++ yierdis-server/src/test/java/yier/bubu/redis/CustomProtocolResyncIntegrationTest.java
@@ -69,6 +69,7 @@ public class CustomProtocolResyncIntegrationTest {
             executor.start();
             this.ch = new EmbeddedChannel(
                     new CustomRequestDecoder(1024 * 1024, 1024, 256),
+                    new ProtocolErrorReplyHandler(executor),
                     new YierdisFastCommandHandler(executor)
             );
         }